backend/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
python-multipart
pydantic
Pillow
google-cloud-storage
